Renamed config.py to common.py because it holds global functions, not just config values. Also moved print_field into common.py
