+++ yaas/common.py
@@ -37,3 +37,21 @@ def href(endpoint):
         server=server,
         port=port,
         path=endpoint)
+
+def print_field(k, v, indent=0):
+    k = str(k).replace('_', ' ')
+    if type(v) is dict:
+        print('{indent}{key}:'.format(indent=' '*indent, key=k))
+        for key, value in v.items():
+            print_field(k=key, v=value, indent=indent+4)
+    elif type(v) is list:
+        print('{indent}{key}:'.format(indent=' '*indent, key=k))
+        for i in range(len(v)):
+            print_field(k=i, v=v[i], indent=indent+4)
+    else:
+        print(
+            '{indent}{key}: {value}'.format(
+                indent=' '*indent,
+                key=k,
+                value=' '.join(v) if type(v) is list else v))
+
